Reverted untechables to vanilla (but kept the spinny)
