+++ src/fighter/common/common_status.rs
@@ -21,8 +21,8 @@ mod damage;
 mod damage_air;
 mod damage_fall;
 mod damage_fly;
-mod damage_fly_roll;
-mod damage_fly_reflect;
+// mod damage_fly_roll;
+// mod damage_fly_reflect;
 mod landing;
 mod pass;
 mod cliff;
@@ -53,8 +53,8 @@ pub fn install() {
     damage_air::install();
     damage_fall::install();
     damage_fly::install();
-    damage_fly_roll::install();
-    damage_fly_reflect::install();
+    // damage_fly_roll::install();
+    // damage_fly_reflect::install();
     landing::install();
     pass::install();
     cliff::install();
